requirements.txt
boxsdk==1.5.3
click==6.6
enum34==1.1.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-http-client==2.2.1
PyYAML==3.12
requests==2.11.1
requests-toolbelt==0.7.0
sendgrid==3.3.1
six==1.10.0
Werkzeug==0.11.11
